changed knots_to_mps to accept a double as a parameter

+++ ELMS/base_station/calculations.cpp
@@ -9,7 +9,7 @@
 knots_to_mps
 =============
 */
-double Calculations::knots_to_mps(int knots){
+double Calculations::knots_to_mps(double knots){
 
 	double mps = -1;
